Beta 1. A few major things left in the TODO list

+++ Assignment6/Forms/FormMain.cs
@@ -8,6 +8,7 @@
 using Assignment6.Classes;
 using Assignment6.Enums;
 using Assignment6.Forms;
+using System.Windows.Forms;
 using static Assignment6.Helpers.EnumHelper;
 
 namespace Assignment6
@@ -36,8 +37,6 @@ namespace Assignment6
         {
             InitializeComponent();
 
-            _taskManager = new TaskManager();
-
             InitGUI();
         }
         #endregion
@@ -45,8 +44,16 @@ namespace Assignment6
         /// <summary>
         /// 
         /// </summary>
-        private void InitGUI()
+        private void InitGUI() //TODO: Make this prettier
         {
+            _taskManager = new TaskManager();
+
+            lstTasks.Items.Clear();
+
+            txtToDo.Text = string.Empty;
+            dateTimePicker1.Value = DateTime.Now;
+            lstTasks.Columns.Clear();
+
             dateTimePicker1.CustomFormat = "yyy-MM-dd    HH:mm:ss";
             dateTimePicker1.MinDate = DateTime.Now;
             lblTimer.Text = GetNowToString("HH:mm:ss");
@@ -58,12 +65,16 @@ namespace Assignment6
 
             lstTasks.View = View.Details;
 
+
             lstTasks.Columns.Add("Date", 100, HorizontalAlignment.Left);
             lstTasks.Columns.Add("Time", 100, HorizontalAlignment.Left);
             lstTasks.Columns.Add("Priority", 150, HorizontalAlignment.Left);
             lstTasks.Columns.Add("Description", 550, HorizontalAlignment.Left);
 
             btnAddTask.Enabled = false;
+
+            UpdateFileMenu();
+            UpdateButtons();
 
         }
         /// <summary>
@@ -96,8 +107,7 @@ namespace Assignment6
         /// <param name="e"></param>
         private void btnAddTask_Click(object sender, EventArgs e)
         {
-            //TODO: Null checks and so on...
-            Id id = new Id(Guid.NewGuid());
+            Id id = new Id();
             Date date = new Date(dateTimePicker1.Value.Date);
             Time time = new Time(dateTimePicker1.Value.TimeOfDay);
             Priority priority = new Priority((PriorityType)cmbPriority.SelectedIndex);
@@ -108,6 +118,8 @@ namespace Assignment6
             TaskManager.AddNew(task);
 
             UpdateListView();
+            UpdateFileMenu();
+            UpdateButtons();
         }
         /// <summary>
         /// 
@@ -117,10 +129,10 @@ namespace Assignment6
             lstTasks.Items.Clear();
             foreach (Assignment6.Classes.Task task in TaskManager.TaskList)
             {
-                ListViewItem item = new ListViewItem(task.Date.ToString()); // Create a new ListViewItem with the text for the first column
-                item.SubItems.Add(task.Time.ToString()); // Add a subitem for the second column
-                item.SubItems.Add(task.Priority.ToString()); // Add a subitem for the third column
-                item.SubItems.Add(task.Description.ToString()); // Add a subitem for the fourth column
+                ListViewItem item = new ListViewItem(task.Date.ToString());
+                item.SubItems.Add(task.Time.ToString());
+                item.SubItems.Add(task.Priority.ToString());
+                item.SubItems.Add(task.Description.ToString());
 
                 //lstTasks.Items.Add(item); // Add the ListViewItem to the ListView
 
@@ -184,7 +196,15 @@ namespace Assignment6
         /// <param name="e"></param>
         private void exitToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            this.Close();
+            DialogResult result = MessageBox.Show("Are you sure you want to close application?\n\nAll unsaved data will be lost!", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
+            {
+                Application.Exit();
+            }
+            else if (result == DialogResult.No)
+            {
+                return; //Do nothing
+            }
         }
         /// <summary>
         /// 
@@ -216,10 +236,43 @@ namespace Assignment6
             {
                 TaskManager.Change(formEdit.TaskCopy);
                 UpdateListView();
+                UpdateFileMenu();
+                UpdateButtons();
+
             }
             else
             {
                 //Nothing for now...
+            }
+        }
+        /// <summary>
+        /// 
+        /// </summary>
+        private void UpdateFileMenu()
+        {
+            if (lstTasks.Items.Count > 0)
+            {
+                saveDataFileToolStripMenuItem.Enabled = true;
+            }
+            else
+            {
+                saveDataFileToolStripMenuItem.Enabled = false;
+            }
+        }
+        /// <summary>
+        /// 
+        /// </summary>
+        private void UpdateButtons()
+        {
+            if (lstTasks.Items.Count > 0)
+            {
+                btnChange.Enabled = true;
+                btnDelete.Enabled = true;
+            }
+            else
+            {
+                btnChange.Enabled = false;
+                btnDelete.Enabled = false;
             }
         }
         /// <summary>
@@ -266,9 +319,20 @@ namespace Assignment6
                 return;
             }
 
-            TaskManager.Delete(id);
 
-            UpdateListView();
+            DialogResult result = MessageBox.Show("Are you sure you want to delete task?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
+            {
+                TaskManager.Delete(id);
+
+                UpdateListView();
+                UpdateFileMenu();
+                UpdateButtons();
+            }
+            else if (result == DialogResult.No)
+            {
+                return; //Do nothing
+            }
         }
         /// <summary>
         /// 
@@ -277,12 +341,37 @@ namespace Assignment6
         /// <param name="e"></param>
         private void saveDataFileToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            Classes.File file = new Classes.File(@"C:\files", "SaveFile.txt");
-            TaskManager.SaveToFile(file);
+            try
+            {
+                SaveFileDialog saveFileDialog = new SaveFileDialog();
+                saveFileDialog.Filter = "Text files (*.txt)|*.txt";
 
-            //Json method
-            //Classes.File file2 = new Classes.File(@"C:\files", "save1.json");
-            //TaskManager.SaveToJsonFile(file2);
+                if (saveFileDialog.ShowDialog() == DialogResult.OK)
+                {
+                    string filePath = saveFileDialog.FileName;
+
+                    if (string.IsNullOrWhiteSpace(filePath))
+                        return;
+
+                    string path = Path.GetDirectoryName(filePath);
+                    string fileName = Path.GetFileName(filePath);
+
+                    Classes.File file = new Classes.File(path, fileName);
+                    TaskManager.SaveToFile(file);
+                }
+                else
+                {
+                    //No need to do anything...
+                }
+
+                //Json method
+                //Classes.File file2 = new Classes.File(@"C:\files", "save1.json");
+                //TaskManager.SaveToJsonFile(file2);
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message, "Error");
+            }
         }
         /// <summary>
         /// 
@@ -291,14 +380,74 @@ namespace Assignment6
         /// <param name="e"></param>
         private void openDataFileToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            Classes.File file = new Classes.File(@"C:\files", "SaveFile.txt");
-            TaskManager.ReadFromFile(file);
-            UpdateListView();
+            try
+            {
+                OpenFileDialog openFileDialog = new OpenFileDialog();
+                openFileDialog.Filter = "Text files (*.txt)|*.txt";
 
-            //Json method
-            //Classes.File file2 = new Classes.File(@"C:\files", "save1.json");
-            //TaskManager.ReadFromJsonFile(file2);
-            //UpdateListView();
+                if (openFileDialog.ShowDialog() == DialogResult.OK)
+                {
+                    string filePath = openFileDialog.FileName;
+
+                    if (string.IsNullOrWhiteSpace(filePath))
+                        return;
+
+                    string path = Path.GetDirectoryName(filePath);
+                    string fileName = Path.GetFileName(filePath);
+
+                    Classes.File file = new Classes.File(path, fileName);
+                    TaskManager.ReadFromFile(file);
+                    UpdateListView();
+                }
+                else
+                {
+                    //No need to do anything...
+                }
+
+                //Json method
+                //Classes.File file2 = new Classes.File(@"C:\files", "save1.json");
+                //TaskManager.ReadFromJsonFile(file2);
+                //UpdateListView();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message, "Error");
+            }
+        }
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void newToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            DialogResult result = MessageBox.Show("Are you sure you want to reload application?\n\nAll unsaved data will be lost!", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
+            {
+                InitGUI();
+            }
+            else if (result == DialogResult.No)
+            {
+                return; //Do nothing
+            }
+        }
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void lstTasks_DoubleClick(object sender, EventArgs e)
+        {
+            btnChange_Click(sender, e);
+        }
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void timer2_Tick(object sender, EventArgs e)
+        {
+            dateTimePicker1.MinDate = DateTime.Now;
         }
     }
 }
